Adds a bunch of stuff
Implements a half-hour cooldown on the callout post. This is likely to
change to be admin configurable on a per-server basis when I get around
to it.

The bot no longer uses the forbidden word to prevent weird interactions
with other bots and things.

Admins now have access to a set of commands to squelch and unsquelch the
bot.

All users may now access a help command to learn what is available.