Better Key Usage
The Bot's key is now stored in an external text file. This should make
updating a little easier, and less prone to accidentally sharing the
key. This might also allow the script to run multiple bots at once, but
I don't know how that would work and also it's bad.