Implements basic sharding, Fixes Key file issue
It worked in my test, but I have no real way to check until testing with
the real bot.

Generates a number of shards based off the number of servers listed in
the timestamps file (there might be a better way to do this.)

In addition, this should fix the issue where whitespace could mess up
the parsing of keys from the key file, which caused issues on Linux.