Implements infrastructure for customTimeouts
Adds a new command, "!vtct" that returns the length of the currently
defined timeout, as well as how long it will be until the next alert may
be issued.

To this end, a new column has been added to the database which denotes
the length, in seconds, of that server's timeout length.