Updates Bot to use our API instead of a local DB (#52)
* Updates Bot to use our API instead of a local DB

This is part of microservice-izing the bot, to allow for building the
dashboard and other such things. Needs lots of re-factoring out of the
"server dao" paradigm.

* Refactors from ServerDao and dicts to classes

A new Server class has been added to handle its functionality, as has a
Ban class. This intends to later factor out a significant portion of bot.py