Fixes At-mentions not detecting correctly
The word boundary check was causing things to fail if the banned word
was book-ended by special characters. This should be fixed now.

However, a known limitation of this is that it won't match correctly if
there are multiple special characters in a row in the infringing
message, like: "<@12345>, test"