Request limiting (#62)
* Adds Session to ban and server

This should, eventually, help cut down on server issues, since bans
should now be able to take a session object to make requests, which will
give a persistent connection. The same is true of servers, which pass
their session to bans now.

* Makes persistant session

This is pretty rad, since now I don't need to add the authorization
header everywhere. The session is created as part of main, then passed
through to everything. If/When the bot is made a class, it should be a
property of it instead of passing it around.

* Updates Server link

Linking to the rules page is a little silly. I link to the about page
now.

* Updates the session object

Now it'll retry and exponentially back-off, which will hopefully help
some things out, maybe.

* Updates tests

Whoops, I broke a bunch of tests