Updates vt and vtct to take optional parameters (#70)
* Updates VTCT to take an optional argument

The command may now be called with an optional index. If the index is
valid (above 0, less than the number of banned worse), the command will
only return results for the specified ban.

If the index is invalid, or absent, information will be returned for all
banned words.

* Adds optional parameter to VT

Similar to the addition to !vtct, this command will now accept an
optional index parameter. If valid, only that particular word will be
returned.