Builds test environment
Tests can now be run using a container. the `tester` make target
has been added for this purpose.
To run tests:
1) Ensure you have a banned-word-tracker image locally
2) run `make tester`
3) run `make test`

A volume is created tying your working directory in the host to the
container's working directory, so tests should be updated as
appropriate. In addition, tester overrides the expected environment
and entrypoint with bash and a set of test credentials.