Dockerize (#92)
* Creates Dockerfile

The First step on properly dockerizing things,
we should now be able to build images that will run the bot.
To this end, I also had to update the scripts to not buffer output, so
I could actually see logs.

* Updates Bot to use envvars instead

Needed for eventual continuous deployment, and needed for dockerizing
things.

* Creates DockerCompose file

Sets up a way to start the needed services to test locally.
Still need to set up make targets to manage things, and a way
to run the unit tests nicely.

* Builds test environment

Tests can now be run using a container. the `tester` make target
has been added for this purpose.
To run tests:
1) Ensure you have a banned-word-tracker image locally
2) run `make tester`
3) run `make test`

A volume is created tying your working directory in the host to the
container's working directory, so tests should be updated as
appropriate. In addition, tester overrides the expected environment
and entrypoint with bash and a set of test credentials.

* Fixes local db

Whoopsie, it's important to bind the db as a volume, so it can actually
persist.

Co-authored-by: matthew-robertson <m33rober@uwaterloo.ca>