Introduces the record command
I don't feel happy about having to change all these tests to support
the addition of an extra parameter on a ban. This should be worked out
into something cleaner.

Testing also needs to be reworked on commands in some way, updating
everything to make specific text is a nightmare.