Updates DiscordPy to version 2
This is important not only for work I'd like to do going forward,
but also just generally essential for the bot to keep working past the
end of August. This commit migrates to DiscordPy Version 2, and ensures
we have the message content intent.

Later work will be done to migrate to slash commands. There isn't much
of a rush here since we have the message intent anyway.